setup(test): dedecated test folder, split unit and jsdom tests
