+++ components/editor/utils/createBlockButton.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import test from 'tape'
 import spy from 'spy'
-import { shallow } from '../../../lib/utils/enzyme'
+import { shallow } from '../../../test/utils/enzyme'
 import createBlockButton from './createBlockButton'
 import { Value } from 'slate'
