+++ components/editor/utils/createPropertyForm.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import test from 'tape'
-import { shallow } from '../../../lib/utils/enzyme'
+import { shallow } from '../../../test/utils/enzyme'
 import createPropertyForm from './createPropertyForm'
 
 test('utils.createPropertyForm', assert => {
