+++ package.json
@@ -12,10 +12,12 @@
     "build": "next build",
     "heroku-postbuild": "npm run build",
     "tape": "cross-env NODE_ENV=test babel-tape-runner",
-    "test": "cross-env NODE_ENV=test babel-tape-runner lib/utils/jsdom.js \"?(lib|pages|components)/**/*.test.js\" | faucet",
+    "test": "npm run test:unit && npm run test:jsdom",
+    "test:unit": "cross-env NODE_ENV=test babel-tape-runner \"?(lib|pages|components)/**/*.test.js\" | faucet",
+    "test:jsdom": "cross-env NODE_ENV=test babel-tape-runner test/jsdom/setup.js \"test/jsdom/**/*.test.js\" | faucet",
     "cloc": "find . -name '*.js' -not -path \"./node_modules*\" -not -path \"./.next*\" | xargs wc -l",
     "standardize": "standard --fix",
-    "precommit": "npm test && lint-staged",
+    "precommit": "npm run test:unit && lint-staged",
     "translations": "gsheets --key=1qPubCKz2YzbcFkRfF0MrgRqD0KLsgRPxoZfG5G3tHOA --title=live --pretty --out lib/translations.json",
     "preinstall": "rm -rf .next && rm -rf node_modules/glamor && rm -rf node_modules/@project-r/styleguide",
     "link:sg": "rm -rf .next && npm link @project-r/styleguide && rm -rf node_modules/glamor && ln -s @project-r/styleguide/node_modules/glamor node_modules/glamor && echo '\n⚠️  Make sure to restart the next.js server. To unlink simply run npm install.\n'"
